refactor(adoc-core): unit 2 — relocate pure domain leaves under domain/
git mv identity.rs, source.rs, diagnostic.rs, inline.rs, ast.rs into
src/domain/. Declare them in domain/mod.rs; remove from lib.rs.
Sweep all callers (compile.rs, parser/, render/, artifact/, validate.rs,
scan/, source_provider.rs) to use crate::domain::* paths.

Refs #27.
