+++ crates/adoc-core/src/domain/diagnostic.rs
