+++ crates/adoc-core/src/domain/mod.rs
@@ -1,1 +1,6 @@
 // see ADR-0009
+pub(crate) mod ast;
+pub(crate) mod diagnostic;
+pub(crate) mod identity;
+pub(crate) mod inline;
+pub(crate) mod source;
