+++ crates/adoc-core/src/lib.rs
@@ -1,19 +1,14 @@
 mod application;
 mod artifact;
-mod ast;
 mod compile;
-mod diagnostic;
 mod domain;
-mod identity;
 mod infrastructure;
-mod inline;
 mod parser;
 mod render;
 mod scan;
-mod source;
 mod source_provider;
 mod validate;
 
 pub use artifact::agent_json::AgentJsonDocument;
 pub use compile::{BuildArtifacts, CompileInput, CompileResult, compile_workspace};
-pub use diagnostic::{Diagnostic, DiagnosticCode, Severity};
+pub use domain::diagnostic::{Diagnostic, DiagnosticCode, Severity};
